conditionally switch plan button text based on user state
Disappointingly, it seems that fetching the user's state from the
supabase cookie fails on the first pass, so the user is stuck on the
pricing page after the redirect as if they weren't logged in. I feel
this may be the cause of the delayed home page change as well.