remove special character from new api keys
By including the > symbol in the api key, it causes issues when used
with the ttm cli unless wrapped in quotes. The experience is less than ideal
for the user, as no error is created (the symbol is intercepted by the terminal
itself and just redirects output to a file).